refactor(scheduler): dedupe destinations + slack hooks, check perms first
Address follow-up review feedback:

- Extract VALID_SCHEDULER_RUN_DESTINATIONS into a single shared module
  (controllers/schedulerConstants.ts), used by the v1 schedulerController
  and both v2 resource-scoped controllers.
- DashboardService/SavedChartService.getSchedulerRuns: run the manage
  permission check before the cross-resource (scheduler ↔ dashboard/chart)
  match, so unauthorized callers can't distinguish 404 from 403 (OWASP
  IDOR concern).
- Extract Slack channel-name resolution into a shared
  useGetSlackChannelName hook, replacing the duplicated
  useGetSlack + useSlackChannels + slackChannelMap setup in
  SchedulerModal, SchedulerRunsHistoryModal, and SchedulersView.

Refs: SPK-449
