+++ packages/frontend/src/components/SchedulersView/index.tsx
@@ -9,9 +9,9 @@ import {
 } from '@mantine-8/core';
 import { IconClock, IconRefresh, IconSend } from '@tabler/icons-react';
 import { useQueryClient } from '@tanstack/react-query';
-import { useCallback, useMemo, useState, type FC } from 'react';
+import { useState, type FC } from 'react';
 import { useSearchParams } from 'react-router';
-import { useGetSlack, useSlackChannels } from '../../hooks/slack/useSlack';
+import { useGetSlackChannelName } from '../../hooks/slack/useGetSlackChannelName';
 import useToaster from '../../hooks/toaster/useToaster';
 import MantineIcon from '../common/MantineIcon';
 import LogsTable from './LogsTable';
@@ -32,45 +32,13 @@ const SchedulersView: FC<{ projectUuid?: string; isUserScope?: boolean }> = ({
             ? SchedulersViewTab.RUN_HISTORY
             : SchedulersViewTab.ALL_SCHEDULERS;
 
-    const {
-        data: slackInstallation,
-        isInitialLoading: isLoadingSlackInstallation,
-    } = useGetSlack();
-    const organizationHasSlack = !!slackInstallation?.organizationUuid;
-
     // Track slack channel IDs from scheduler data to ensure they're included in the query
     const [schedulerSlackChannelIds, setSchedulerSlackChannelIds] = useState<
         string[]
     >([]);
-
-    const slackChannelsQuery = useSlackChannels(
-        '',
-        {
-            excludeArchived: false,
-            includeChannelIds:
-                schedulerSlackChannelIds.length > 0
-                    ? schedulerSlackChannelIds
-                    : undefined,
-        },
-        { enabled: organizationHasSlack && !isLoadingSlackInstallation },
-    );
-
-    // Create a map of Slack channel ID -> name
-    const slackChannelMap = useMemo(() => {
-        const map = new Map<string, string>();
-        slackChannelsQuery?.data?.forEach((channel) => {
-            map.set(channel.id, channel.name);
-        });
-        return map;
-    }, [slackChannelsQuery?.data]);
-
-    // Callback to get Slack channel name from ID
-    const getSlackChannelName = useCallback(
-        (channelId: string): string | null => {
-            return slackChannelMap.get(channelId) || null;
-        },
-        [slackChannelMap],
-    );
+    const { getSlackChannelName } = useGetSlackChannelName({
+        includeChannelIds: schedulerSlackChannelIds,
+    });
 
     const handleTabChange = (value: string | null) => {
         const newParams = new URLSearchParams(searchParams);
